Labelbox-project-setup/requirements.txt
opencv-python
numpy
labelbox
pandas
youtube-dl
python-dotenv